Rename map_err to map_error
The latter is more consistent with the rest of my functions. I think the
method on `Result` in the standard library is called `map_err` because
it maps if the result is `Err(..)`, so in that sense it is not
inconsistent to name my method `map_error`, because there is no `Err`
variant in my case.